return no content if cannot find book to delete

+++ src/Recipes.API/Controllers/RecipesController.cs
@@ -60,6 +60,10 @@ namespace Recipes.API.Controllers
         [HttpDelete("{id:guid}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var recipe = _catalogService.GetRecipe(id);
+            if (recipe == null)
+                return NotFound();
+
             await _catalogService.DeleteRecipe(id);
 
             return NoContent();
